--- garrator/run.py
@@ -1,2 +1,0 @@
-from app import App
-
